AI-Scheduling-Assistant/requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
airportsdata==20250224
aiter @ file:///install/aiter-0.1.0-py3-none-any.whl#sha256=6225f8a878a75002dac1e090d97ab8c24c8395723db127c8e8b810426caf47fb
alabaster==1.0.0
amdsmi @ file:///install/amdsmi-24.7.1%2B0012a68-py3-none-any.whl#sha256=13e7232d9f9ee6af3a973d4aaf865dc7263ce93d26fed568a0e97731574bf8f6
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
blake3==1.0.5
bleach==6.2.0
blinker==1.4
boto3==1.38.20
botocore==1.38.20
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
cmake==3.31.6
cohere==5.16.1
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.9.4
coverage==7.8.2
cryptography==3.4.8
Cython==3.1.0
dask==2025.5.1
dataclasses-json==0.6.7
datasets==3.6.0
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
deepmerge==2.0
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distributed==2025.5.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docutils==0.21.2
einops==0.8.1
email_validator==2.2.0
eval_type_backport==0.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastavro==1.11.1
fastjsonschema==2.21.1
filelock==3.18.0
flash_attn @ file:///install/flash_attn-2.7.4.post1-cp312-cp312-linux_x86_64.whl#sha256=1c52be517b24310ea2894ff84b2c65f97392a8595f931249e31c5b6d87d12958
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.0
gguf==0.16.3
gitdb==4.0.12
GitPython==0.3.6
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.25.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
groq==0.30.0
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.2
HipMarker==1.0
hiredis==3.1.1
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
humanize==4.12.3
idna==3.10
ihighlight==1.0.0
imagesize==1.4.1
importlib_metadata==8.6.1
iniconfig==2.1.0
inquirerpy==0.3.4
interegular==0.3.3
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
json5==0.12.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_ai==2.31.4
jupyter_ai_magics==2.31.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab-lsp==5.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==23.5.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-text-splitters==0.3.8
langsmith==0.3.45
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libnacl==2.1.0
llguidance==0.7.21
llvmlite==0.44.0
lm-format-enforcer==0.10.11
locket==1.0.0
logfire-api==3.24.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.11.0
mcp-server-time==2025.7.1
mcp-server-weather==0.1.4
mdurl==0.1.2
mistral_common==1.5.5
mistralai==1.9.2
mistune==3.1.3
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nose==1.3.7
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.0
openai==1.79.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-semantic-conventions-ai==0.4.9
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
partial-json-parser==0.2.1.1.post5
peft==0.15.2
pexpect==4.9.0
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.4
pydantic-ai==0.4.2
pydantic-ai-slim==0.4.2
pydantic-evals==0.4.2
pydantic-extra-types==2.10.5
pydantic-graph==0.4.2
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pypi-publisher==0.0.4
pytest==8.3.5
pytest-asyncio==0.26.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML @ file:///install/PyYAML-6.0.2-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl#sha256=80bab7bfc629882493af4aa31a4cfa43a4c57c83813253626916b8c7ada83476
pyzmq==26.4.0
ray==2.44.1
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.6
rocpd @ file:///app/rocmProfileData/rocpd_python
roman-numerals-py==3.1.0
rpds-py==0.25.1
rpdTracer==1.0
rsa==4.9.1
runai-model-streamer==0.11.0
runai-model-streamer-s3==0.11.0
s3transfer==0.12.0
safetensors==0.5.3
scipy==1.15.3
SecretStorage==3.3.1
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==79.0.1
setuptools-scm==8.3.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.7
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tblib==3.1.0
tenacity==8.5.0
tensorizer==2.9.3
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toolz==1.0.0
torch @ file:///install/torch-2.7.0%2Bgitf717b2a-cp312-cp312-linux_x86_64.whl#sha256=c337defece6ecf810988f905883e7b751106ee4c255aacb39960e9498bb6ab6a
torchvision @ file:///install/torchvision-0.21.0%2B7af6987-cp312-cp312-linux_x86_64.whl#sha256=bb2191d82f9bdf2e8abbce982277b0ed0baf6043c86725a1f58707e1b221fb3a
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.2
triton @ file:///install/triton-3.2.0%2Bgite5be006a-cp312-cp312-linux_x86_64.whl#sha256=ab04fdb4700d82eaf2d75cabc848372bb4b7e341bf3a51dbef634d9acf39f013
typer==0.15.4
types-python-dateutil==2.9.0.20250516
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vllm @ file:///install/vllm-0.8.6.dev315%2Bg91a560098.rocm631-cp312-cp312-linux_x86_64.whl#sha256=20beba7197879149b09ddf55c1553747f1f12cdceb9ea135f1b3d0f78af625d0
wadllib==1.3.6
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xgrammar==0.1.19
xxhash==3.5.0
yarl==1.20.0
yes==0.0.1
zict==3.0.0
zipp==3.21.0
zstandard==0.23.0
